+++ infra/common/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.metadata.statistics.collector.ShardingSphereStatisticsCollector
@@ -15,5 +15,5 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.infra.metadata.data.builder.dialect.MySQLShardingSphereDataBuilder
-org.apache.shardingsphere.infra.metadata.data.builder.dialect.PostgreSQLShardingSphereDataBuilder
+org.apache.shardingsphere.infra.metadata.statistics.collector.tables.PgNamespaceTableCollector
+org.apache.shardingsphere.infra.metadata.statistics.collector.tables.PgClassTableCollector
